Move exhaustiveness_problem to repl state test
